Extract fetch functionality
