+++ package.json
@@ -33,7 +33,7 @@
     "@types/node": "^14.14.10",
     "@types/rdf-dataset-ext": "^1.0.0",
     "@types/rdf-ext": "^1.3.8",
-    "@types/rdf-js": "^4.0.0",
+    "@types/rdf-js": "^3.0.3",
     "@types/rdfjs__dataset": "^1.0.4",
     "gts": "^3.0.3",
     "ts-node-dev": "^1.1.1",
